合并作者更新

+++ BDW.sh
@@ -2,7 +2,7 @@
 PATH=/bin:/sbin:/usr/bin:/usr/sbin:/usr/local/bin:/usr/local/sbin:~/bin
 export PATH
 
-sh_ver="1.0.1"
+sh_ver="1.0.8"
 file="/root/BaiduPCSWeb"
 Folder="/usr/local/BaiduPCSWeb"
 BaiduPCS_Go="/usr/bin/BaiduPCS-Go"
@@ -23,23 +23,23 @@ Error="${Red_font_prefix}[错误]${Font_color_suffix}"
 Tip="${Green_font_prefix}[注意]${Font_color_suffix}"
 
 check_root(){
-	[[ $EUID != 0 ]] && echo -e "${Error} 当前非ROOT账号(或没有ROOT权限)，无法继续操作，请更换ROOT账号或使用 ${Green_background_prefix}sudo su${Font_color_suffix} 命令获取临时ROOT权限（执行后可能会提示输入当前账号的密码）。" && exit 1
+	[[ $EUID != 0 ]] && echo -ne "${Error} 当前非ROOT账号(或没有ROOT权限)，无法继续操作，请更换ROOT账号或使用 ${Green_background_prefix}sudo su${Font_color_suffix} 命令获取临时ROOT权限（执行后可能会提示输入当前账号的密码）。" && exit 1
 }
 #检查系统
 check_sys(){
 	if [[ -f /etc/redhat-release ]]; then
 		release="centos"
-	elif cat /etc/issue | grep -q -E -i "debian"; then
+	elif cat /etc/issue | grep -qEi "debian"; then
 		release="debian"
-	elif cat /etc/issue | grep -q -E -i "ubuntu"; then
+	elif cat /etc/issue | grep -qEi "ubuntu"; then
 		release="ubuntu"
-	elif cat /etc/issue | grep -q -E -i "centos|red hat|redhat"; then
+	elif cat /etc/issue | grep -qEi "centos|red hat|redhat"; then
 		release="centos"
-	elif cat /proc/version | grep -q -E -i "debian"; then
+	elif cat /proc/version | grep -qEi "debian"; then
 		release="debian"
-	elif cat /proc/version | grep -q -E -i "ubuntu"; then
+	elif cat /proc/version | grep -qEi "ubuntu"; then
 		release="ubuntu"
-	elif cat /proc/version | grep -q -E -i "centos|red hat|redhat"; then
+	elif cat /proc/version | grep -qEi "centos|red hat|redhat"; then
 		release="centos"
     fi
 	bit=`uname -m`
@@ -48,7 +48,7 @@ check_installed_status(){
 	[[ ! -e ${Folder} ]] && echo -e "${Error} BaiduPCS-Web 没有安装，请检查 !" && exit 1
 }
 check_pid(){
-	PID=`ps -ef| grep "BaiduPCS-Go"| grep -v grep| grep -v "BDW.sh"| grep -v "init.d"| grep -v "service"| awk '{print $2}'`
+	PID=$(ps -ef| grep "BaiduPCS-Go"| grep -v grep| grep -v "BDW.sh"| grep -v "init.d"| grep -v "service"| awk '{print $2}')
 }
 
 Set_BaiduPCS_port(){
@@ -61,6 +61,7 @@ Set_BaiduPCS_port(){
 		$BaiduPCS_port)
 			echo
 			echo " 哎呀...跟当前端口一毛一样呀...修改个鸡鸡哦"
+			break
 			;;
 		[1-9] | [1-9][0-9] | [1-9][0-9][0-9] | [1-9][0-9][0-9][0-9] | [1-5][0-9][0-9][0-9][0-9] | 6[0-4][0-9][0-9][0-9] | 65[0-4][0-9][0-9] | 655[0-3][0-5])
 			echo "$BaiduPCS_port_opt" > ${Folder}/port
@@ -94,7 +95,7 @@ check_new_ver(){
 	fi
 }
 check_ver_comparison(){
-	BaiduPCS_Web_now_ver=$(${BaiduPCS_Go} -v|head -n 1|awk '{print $3}')
+	BaiduPCS_Web_now_ver=$(${Folder}/BaiduPCS-Go -v|head -n 1|awk '{print $3}')
 	[[ -z ${BaiduPCS_Web_now_ver} ]] && echo -e "${Error} BaiduPCS_Web 当前版本获取失败 !" && exit 1
 	if [[ "${BaiduPCS_Web_now_ver}" != "${BaiduPCS_Web_new_ver}" ]]; then
 		echo -e "${Info} 发现 BaiduPCS-Web 已有新版本 [ ${BaiduPCS_Web_new_ver} ](当前版本：${BaiduPCS_Web_now_ver})"
@@ -103,18 +104,37 @@ check_ver_comparison(){
 		if [[ $yn == [Yy] ]]; then
 			check_pid
 			[[ ! -z $PID ]] && kill -9 ${PID}
-			Download_BaiduPCS_Web "update"
-			Start_BaiduPCS_Web
+			Download_BaiduPCS_Web
 		fi
 	else
 		echo -e "${Info} 当前 BaiduPCS-Web 已是最新版本 [ ${BaiduPCS_Web_new_ver} ]" && exit 1
 	fi
 }
 
+check_ver_comparison_fix(){
+	BaiduPCS_Web_now_ver=$(${Folder}/BaiduPCS-Go -v|head -n 1|awk '{print $3}')
+	[[ -z ${BaiduPCS_Web_now_ver} ]] && echo -e "${Error} BaiduPCS_Web 当前版本获取失败 !" && exit 1
+	if [[ "${BaiduPCS_Web_now_ver}" != "${BaiduPCS_Web_new_ver}" ]]; then
+		echo -e "${Info} 发现 BaiduPCS-Web 已有新版本 [ ${BaiduPCS_Web_new_ver} ](当前版本：${BaiduPCS_Web_now_ver})"
+		read -e -p "是否更新(会中断当前下载任务，请注意) ? [Y/n] :" yn
+		[[ -z "${yn}" ]] && yn="y"
+		if [[ $yn == [Yy] ]]; then
+			check_pid
+			[[ ! -z $PID ]] && kill -9 ${PID}
+			Download_BaiduPCS_Web
+		fi
+	else
+		echo -e "${Info} 正在重新下载最新版本 [ ${BaiduPCS_Web_new_ver} ]"
+		check_pid
+		[[ ! -z $PID ]] && kill -9 ${PID}
+		Download_BaiduPCS_Web
+	fi
+}
+
 Download_BaiduPCS_Web(){
-	update_dl=$1
 	cd "/usr/local"
 	#echo -e "${bit}"
+	bit=`uname -m`
 	if [[ ${bit} == "x86_64" ]]; then
 		bit="amd64"
 	elif [[ ${bit} == "i386" || ${bit} == "i686" ]]; then
@@ -127,10 +147,15 @@ Download_BaiduPCS_Web(){
 	[[ ! -s ${BaiduPCS_Web_Name}.zip ]] && echo -e "${Error} BaiduPCS-Web 压缩包下载失败 !" && exit 1
 	unzip ${BaiduPCS_Web_Name}.zip
 	[[ ! -e "/usr/local/${BaiduPCS_Web_Name}" ]] && echo -e "${Error} BaiduPCS-Web 解压失败 !" && rm -rf ${BaiduPCS_Web_Name}.zip && exit 1
-	[[ ${update_dl} = "update" ]] && rm -rf "${Folder}"
-	mv "/usr/local/${BaiduPCS_Web_Name}" "${Folder}"
-	[[ ! -e "${Folder}" ]] && echo -e "${Error} BaiduPCS-Web 文件夹重命名失败 !" && rm -rf ${BaiduPCS_Web_Name}.zip && rm -rf "/usr/local/${BaiduPCS_Web_Name}" && exit 1
-	rm -rf ${BaiduPCS_Web_Name}.zip
+	rm -rf "${Folder}/BaiduPCS-Go"
+	cp -f "/usr/local/${BaiduPCS_Web_Name}/BaiduPCS-Go" "${Folder}/BaiduPCS-Go"
+	if [[ ! -e "${Folder}" ]]; then 
+		echo -e "${Error} ${Folder}/BaiduPCS-Go 复制失败 !首次安装请忽略此信息！" 
+		mv "/usr/local/${BaiduPCS_Web_Name}" "${Folder}"	
+	fi
+	[[ ! -e "${Folder}" ]] && echo -e "${Error} BaiduPCS-Web 文件夹重命名失败 !" && rm -rf "/usr/local/${BaiduPCS_Web_Name}.zip" && rm -rf "/usr/local/${BaiduPCS_Web_Name}" && exit 1
+	rm -rf "/usr/local/${BaiduPCS_Web_Name}.zip"
+	rm -rf "/usr/local/${BaiduPCS_Web_Name}"
 	cd "${Folder}"
 	chmod a+x BaiduPCS-Go
 	echo -e "${Info} BaiduPCS-Web 主程序安装完毕！..."
@@ -158,7 +183,7 @@ Service_BaiduPCS_Web(){
 }
 
 Download_BaiduPCS_port(){
-	if ! wget --no-check-certificate https://raw.githubusercontent.com/user1121114685/baidupcsweb/master/port -O ${Folder}/port; then
+	if ! wget --no-check-certificate https://raw.githubusercontent.com/user1121114685/baidupcsweb/master/port -O "${Folder}/port"; then
 		echo -e "${Error} BaiduPCS-Web服务 prot下载失败 !" && exit 1
 	fi
 	echo -e "成功下载port文件..."
@@ -176,6 +201,12 @@ Installation_dependency(){
  }
 Install_BaiduPCS_Web(){
 	check_root
+	check_pid
+	[[ ! -z $PID ]] && kill -9 ${PID}
+	rm -rf "${BaiduPCS_Go}"
+	rm -rf "${Folder}"
+	rm -rf "${file}"
+	rm -rf "${Folder}/port"
 	[[ -e ${BaiduPCS_Go} ]] && echo -e "${Error} BaiduPCS-Web 已安装，请检查 !" && exit 1
 	check_sys
 	echo -e "${Info} 开始安装/配置 依赖..."
@@ -213,7 +244,15 @@ Update_BaiduPCS_Web(){
 	check_installed_status
 	check_new_ver
 	check_ver_comparison
+	Start_BaiduPCS_Web
 }
+
+Fix_BaiduPCS_Web(){
+	check_installed_status
+	check_new_ver
+	check_ver_comparison_fix
+}
+
 UnInstall_BaiduPCS_Web(){
 	check_installed_status "un"
 	echo "确定要卸载 BaiduPCS-Web ? (y/N)"
@@ -226,6 +265,7 @@ UnInstall_BaiduPCS_Web(){
 		rm -rf "${BaiduPCS_Go}"
 		rm -rf "${Folder}"
 		rm -rf "${file}"
+		rm -rf "${Folder}/port"
 		if [[ ${release} = "centos" ]]; then
 			chkconfig --del BaiduPCSWeb
 		else
@@ -265,11 +305,49 @@ Update_Shell(){
 	fi
 }
 
+Set_BaiduPCS_Advanced(){
+	echo && echo -e " 你要做什么？
+	
+ ${Green_font_prefix} 0.${Font_color_suffix} 尝试修复（不用重装）
+————————
+ ${Green_font_prefix} 1.${Font_color_suffix} 显示已登录的账户
+ ${Green_font_prefix} 2.${Font_color_suffix} 占位（暂时不知道做什么）
+————————
+ ${Green_font_prefix} 3.${Font_color_suffix} 清空所有账号信息（包括密码与登陆的账号）
+ 
+ 注意：部分操作会自动重启客户端！" && echo
+	read -e -p "(默认: 取消):" set_num
+	[[ -z "${set_num}" ]] && echo "已取消..." && exit 1
+	if [[ ${set_num} == "0" ]]; then
+	echo -e "$yellow 开始修复，请保证你目前使用的是最新脚本。$none"
+	echo -e ""
+	Fix_BaiduPCS_Web
+	Service_BaiduPCS_Web
+	ReStart_BaiduPCS_Web
+
+	elif [[ ${set_num} == "1" ]]; then
+	echo -e "$yellow您已登录以下账户"
+	echo -e ""
+	cat "${Folder}/pcs_config.json" | grep name | cut -d\" -f4
+	echo -e "$yellow $none"
+
+	elif [[ ${set_num} == "3" ]]; then
+	rm -rf "${Folder}/pcs_config.json"
+	echo -e "$yellow 配置文件已经删除，请重新登录账号！$none"
+	ReStart_BaiduPCS_Web
+	else
+		echo -e "${Error} 请输入正确的数字[1-3]" && exit 1
+	fi
+}
+
+
 echo && echo -e " BaiduPCS-Web 一键安装管理脚本 ${Red_font_prefix}[v${sh_ver}]${Font_color_suffix}
 
 		by 联盟少侠 
 
-管理地址:http://你的IP:${BaiduPCS_port}(支持外网访问)
+>>>>>>管理地址:http://你的IP:${BaiduPCS_port}(支持外网访问)<<<<<<<
+
+>>>>>>如果不能访问，请自行进行端口放行，或者端口映射<<<<<<<<<
 
 BaiduPCS-Web项目地址：https://github.com/liuzhuoling2011/baidupcs-web
 
@@ -278,17 +356,18 @@ BaiduPCS-Web项目地址：https://github.com/liuzhuoling2011/baidupcs-web
 
   
  ${Green_font_prefix} 0.${Font_color_suffix} 升级脚本
-————————————
+————————————————————————
  ${Green_font_prefix} 1.${Font_color_suffix} 安装 BaiduPCS-Web
- ${Green_font_prefix} 2.${Font_color_suffix} 更新 BaiduPCS-Web
+ ${Green_font_prefix} 2.${Font_color_suffix} 更新 BaiduPCS-Web(可指定版本)
  ${Green_font_prefix} 3.${Font_color_suffix} 卸载 BaiduPCS-Web
-————————————
+————————————————————————
  ${Green_font_prefix} 4.${Font_color_suffix} 启动 BaiduPCS-Web
  ${Green_font_prefix} 5.${Font_color_suffix} 停止 BaiduPCS-Web
  ${Green_font_prefix} 6.${Font_color_suffix} 重启 BaiduPCS-Web
- ————————————
+————————————————————————
  ${Green_font_prefix} 7.${Font_color_suffix} 修改 BaiduPCS-Web 端口
-————————————" && echo
+ ${Green_font_prefix} 8.${Font_color_suffix} 高级功能
+————————————————————————" && echo
 if [[ -e ${Folder} ]]; then
 	check_pid
 	if [[ ! -z "${PID}" ]]; then
@@ -300,7 +379,7 @@ else
 	echo -e " 当前状态: ${Red_font_prefix}未安装${Font_color_suffix}"
 fi
 echo
-read -e -p " 请输入数字 [0-7]:" num
+read -e -p " 请输入数字 [0-8]:" num
 case "$num" in
 	0)
 	Update_Shell
@@ -326,7 +405,10 @@ case "$num" in
 	7)
 	Set_BaiduPCS_port
 	;;
+	8)
+	Set_BaiduPCS_Advanced
+	;;
 	*)
-	echo "请输入正确数字 [0-7]"
+	echo "请输入正确数字 [0-8]"
 	;;
 esac
